+++ src/App.tsx
@@ -34,7 +34,7 @@ function App() {
           <Route path="/" element={<Layout />}>
             <Route index element={<Home />} />
             <Route path="user" element={<User />} />
-            <Route path="product-detail" element={<ProductDetail />} />
+            <Route path="product-detail" element={<ProductDetail  productId={1}/>} />
             <Route path="login" element={<Login />} />
             <Route path="register" element={<Register />} />
             <Route path="cart" element={<Cart />} />
