+++ src/app/store.ts
@@ -1,10 +1,12 @@
 import { configureStore } from "@reduxjs/toolkit";
 import cartReducer from "../features/cart/cartSlice";
 import wishlistReducer from "../features/wishlist/wishlistSlice"
+import productDetailSlice from "../features/productDetail/productDetailSlice.ts";
 export const store = configureStore({
     reducer: {
         cart: cartReducer,
-        wishlist: wishlistReducer
+        wishlist: wishlistReducer,
+        productDetail: productDetailSlice,
     }
 })
 export type RootState = ReturnType<typeof store.getState>
