+++ src/components/Cart/CartContainer.tsx
@@ -1,12 +1,13 @@
-import { Box, Button, IconButton, Paper, Stack, Typography } from '@mui/material'
+import {Box, Button, Paper, Stack, Typography} from '@mui/material'
 import React from 'react'
-import { useDispatch, useSelector } from 'react-redux'
-import { RootState } from '../../app/store'
-import { VNDNumericFormat } from '../ProductCard'
-import { CartItem, cartItemRemove, getTotalPrice } from '../../features/cart/cartSlice'
+import {useDispatch, useSelector} from 'react-redux'
+import {RootState} from '../../app/store'
+import {VNDNumericFormat} from '../ProductCard'
+import {CartItem, cartItemRemove, getTotalPrice} from '../../features/cart/cartSlice'
 import CancelIcon from '@mui/icons-material/Cancel';
-import { toast } from 'react-toastify'
-import { useNavigate } from 'react-router-dom'
+import {toast} from 'react-toastify'
+import {useNavigate} from 'react-router-dom'
+
 type CartItemListType = {
     cartItems: CartItem[];
 }
