+++ src/features/productDetail/review.ts
@@ -1,0 +1,5 @@
+import {IReview} from "../../interfaces/IProductDescription.ts";
+import {productDetails} from "../../data/productDetail.ts";
+
+
+const initialState: IReview[] = productDetails.
