+++ src/interfaces/Customer.ts
@@ -1,0 +1,4 @@
+export interface Customer {
+    id: number;
+    fullName: string;
+}
