+++ src/interfaces/ProductDetail.ts
@@ -1,0 +1,9 @@
+import {Product} from "./Product.ts";
+import {Image, ProductDescription} from "./IProductDescription.ts";
+
+// export interface IProductDetail extends Product{
+//     discount: number;
+//     images: Image[];
+//     overview: string;
+//     productDescription: ProductDescription[];
+// }
